refactor client mod
